requirements.txt
redis
numpy
gradio
numpy
pillow


